# swagger.yaml
openapi: 3.0.3
info:
  title: Favourites API
  description: User favourites (charts, insights, audiences) with JWT auth. API is versioned under the path prefix (e.g. /v1).
  version: 1.0.0
servers:
  - url: /v1
    description: API v1 (current)
paths:
...
  /auth/token:
    post:
      summary: Issue JWT
      operationId: issueToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IssueTokenRequest'
      responses:
        '200':
          description: Token issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssueTokenResponse'
        '400':
          description: Bad request (e.g. missing user_id)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '405':
          description: Method not allowed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /users/{userID}/favourites:
    get:
      summary: List favourites (paginated)
      operationId: listFavourites
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UserID'
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: Paginated list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListFavouritesResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
    post:
      summary: Add favourite
      operationId: addFavourite
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UserID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddFavouriteRequest'
      responses:
        '201':
          description: Favourite created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FavouriteResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
  /users/{userID}/favourites/{assetID}:
    get:
      summary: Get one favourite
      operationId: getFavourite
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UserID'
        - $ref: '#/components/parameters/AssetID'
      responses:
        '200':
          description: Favourite details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FavouriteResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
    delete:
      summary: Remove favourite
      operationId: removeFavourite
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UserID'
        - $ref: '#/components/parameters/AssetID'
      responses:
        '204':
          description: Removed
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
    patch:
      summary: Update favourite description
      operationId: updateFavouriteDescription
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UserID'
        - $ref: '#/components/parameters/AssetID'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateDescriptionRequest'
      responses:
        '200':
          description: Updated favourite
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FavouriteResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
  /internal/events/asset-deleted:
    post:
      summary: Internal event – asset deleted (optional X-Internal-Secret)
      operationId: assetDeleted
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssetDeletedRequest'
      responses:
        '200':
          description: Favourites removed count
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetDeletedResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT from POST /auth/token
  parameters:
    UserID:
      name: userID
      in: path
      required: true
      schema:
        type: string
    AssetID:
      name: assetID
      in: path
      required: true
      schema:
        type: string
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Missing or invalid token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Cannot access another user's favourites
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    MethodNotAllowed:
      description: Method not allowed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    IssueTokenRequest:
      type: object
      required: [user_id]
      properties:
        user_id:
          type: string
    IssueTokenResponse:
      type: object
      properties:
        token:
          type: string
        expires_in:
          type: integer
          description: Seconds until expiry
    FavouriteListItem:
      type: object
      properties:
        id: { type: string }
        type: { type: string }
        description: { type: string }
        source_asset_id: { type: string }
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ListFavouritesResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/FavouriteListItem'
        total: { type: integer }
        limit: { type: integer }
        offset: { type: integer }
    ChartData:
      type: object
      properties:
        title: { type: string }
        x_axis_title: { type: string }
        y_axis_title: { type: string }
        data: {}
    InsightData:
      type: object
      properties:
        text: { type: string }
    AudienceData:
      type: object
      properties:
        gender: { type: string }
        birth_country: { type: string }
        age_groups: { type: string }
        hours_social_media_daily: { type: number }
        purchases_last_month: { type: integer }
    FavouriteResponse:
      type: object
      properties:
        id: { type: string }
        type: { type: string }
        description: { type: string }
        source_asset_id: { type: string }
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        chart: { $ref: '#/components/schemas/ChartData' }
        insight: { $ref: '#/components/schemas/InsightData' }
        audience: { $ref: '#/components/schemas/AudienceData' }
    AddFavouriteRequest:
      type: object
      required: [type]
      properties:
        type: { type: string }
        description: { type: string }
        source_asset_id: { type: string }
        chart: { $ref: '#/components/schemas/ChartData' }
        insight: { $ref: '#/components/schemas/InsightData' }
        audience: { $ref: '#/components/schemas/AudienceData' }
    UpdateDescriptionRequest:
      type: object
      properties:
        description: { type: string }
    AssetDeletedRequest:
      type: object
      required: [asset_id]
      properties:
        asset_id: { type: string }
    AssetDeletedResponse:
      type: object
      properties:
        removed: { type: integer }
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string }
            message: { type: string }